update cond image transform - ints instead of floats

+++ costar_models/python/costar_models/multi_sampler.py
@@ -341,7 +341,7 @@ class RobotMultiPredictionSampler(RobotMultiHierarchical):
             x = AddDense(x, 256, "relu", 0.,
                     constraint=None, output=False,)
             x = AddDense(x, int(h_dim[0] * h_dim[1] * 32/4), "relu", 0., constraint=None, output=False)
-            x = Reshape([h_dim[0]/2, h_dim[1]/2, 32])(x)
+            x = Reshape([int(h_dim[0]/2), int(h_dim[1]/2), 32])(x)
         else:
             x = AddConv2D(x, 128, [5,5], 1, 0.)
         x = AddConv2DTranspose(x, 64, [5,5], 2,
